Refactored MAAT API client classes to adapt to suggested repository crime-apps-adaptor
