--- dces-report-service/src/main/java/uk/gov/justice/laa/crime/dces/report/maatapi/config/RetryConfiguration.java
@@ -1,21 +1,0 @@
-package uk.gov.justice.laa.crime.dces.report.maatapi.config;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-
-@Data
-@Configuration
-@ConfigurationProperties(prefix = "maat-retry-config")
-public class RetryConfiguration {
-    @NotNull
-    private Integer maxRetries;
-
-    @NotNull
-    private Integer minBackOffPeriod;
-
-    @NotNull
-    private Double jitterValue;
-}
