+++ dces-report-service/src/main/java/uk/gov/justice/laa/crime/dces/report/maatapi/resources/application.yaml
@@ -3,11 +3,6 @@ maatApi:
   oAuthEnabled: true
   baseUrl: ${MAAT_API_BASE_URL}
 
-maat-retry-config:
-  max-retries: 2
-  min-back-off-period: 3
-  jitter-value: 0.5
-
 means-assessment:
   security:
     issuer-uri: http://localhost:${server.port}
